move wifi setup page to client itself
