+++ src/implementations/the_element/lib/wifi.js
@@ -1,13 +1,9 @@
 'use strict';
 
-const exec       = require('child_process').exec;
 const fs         = require('fs');
-const uuid       = require('uuid');
-const Handlebars = require('handlebars');
+const exec       = require('child_process').exec;
 const bashEscape = require('../bashutils').escape;
 const service    = 'sudo fiw'; // NB: all sudo actions have to be permitted in sudoers.d/formide
-
-Handlebars.registerHelper('list', (ctx, opt) => ctx.reduce((prev, curr) => prev + opt.fn(curr), ''));
 
 /**
  * Filter list of networks to only return valid essids
@@ -172,7 +168,7 @@ module.exports = {
     },
 
     /**
-     * Reset Wi-Fi and fall back to hotspot mode
+     * Reset Wi-Fi and fall back to hot-spot mode
      * @param callback
      */
     reset(callback) {
@@ -181,45 +177,6 @@ module.exports = {
                 return callback(err);
 
             return callback(null, { message: "Successfully reset wlan0" });
-        });
-    },
-
-    /**
-     * Get setup page for Wi-Fi and Formide account connection
-     * @param platformUrl
-     * @param callback
-     */
-    getWlanSetupPage(platformUrl, callback) {
-        fs.readFile(__dirname + "/networks.html", 'utf8', (err, data) => {
-            if (err)
-                return callback(err);
-
-            const template = Handlebars.compile(data);
-
-            // wifi.mac((macErr, macAddress) => {
-            //     if (macErr)
-            //         return callback(macErr);
-            //
-            //     wifi.list((wifiErr, ssids) => {
-            //         if (wifiErr)
-            //             return callback(wifiErr);
-            //
-                    const networks = [];
-            //         for (const ssid in ssids)
-            //             networks.push({ ssid });
-
-                    const registrationToken = uuid.v4();
-                    const redirectUrl = `${platformUrl}/#/manage/devices/setup?registration_token=${registrationToken}`
-
-                    const html = template({
-                        networks,
-                        // macAddress,
-                        registrationToken,
-                        redirectUrl
-                    });
-                    return callback(null, html);
-            //     });
-            // });
         });
     }
 };
